prq/draft/django-rest-framework/requirements.txt
requests==2.4.1
rq==0.4.6
rq-dashboard==0.3.4
nose
django
djangorestframework
markdown
django-filter